+++ _integrations/nrlogrus/nrlogrus.go
@@ -1,14 +1,23 @@
-// Package nrlogrus forwards go-agent log messages to logrus.  If you are using
-// logrus for your application and would like the go-agent log messages to end
-// up in the same place, modify your config as follows:
+// Package nrlogrus sends go-agent log messages to
+// https://github.com/sirupsen/logrus.
 //
-//    cfg.Logger = nrlogrus.StandardLogger()
+// Use this package if you are using logrus in your application and would like
+// the go-agent log messages to end up in the same place.  If you are using
+// the logrus standard logger, assign the newrelic.Config.Logger field to
+// nrlogrus.StandardLogger():
 //
-// Only logrus' StandardLogger is supported since there is no method (as of July
-// 2016) to get the level of a logrus.Logger. See
-// https://github.com/sirupsen/logrus/issues/241
+//	cfg := newrelic.NewConfig("Your Application Name", "__YOUR_NEW_RELIC_LICENSE_KEY__")
+//	cfg.Logger = nrlogrus.StandardLogger()
 //
-// Example: https://github.com/newrelic/go-agent/tree/master/_integrations/nrlogrus/example/main.go
+// If you are using a particular logrus Logger instance, assign the
+// newrelic.Config.Logger field to the the output of nrlogrus.Transform:
+//
+//	l := logrus.New()
+//	l.SetLevel(logrus.DebugLevel)
+//	cfg := newrelic.NewConfig("Your Application Name", "__YOUR_NEW_RELIC_LICENSE_KEY__")
+//	cfg.Logger = nrlogrus.Transform(l)
+//
+// This package requires logrus version v1.1.0 and above.
 package nrlogrus
 
 import (
@@ -19,7 +28,10 @@ import (
 
 func init() { internal.TrackUsage("integration", "logging", "logrus") }
 
-type shim struct{ e *logrus.Entry }
+type shim struct {
+	e *logrus.Entry
+	l *logrus.Logger
+}
 
 func (s *shim) Error(msg string, c map[string]interface{}) {
 	s.e.WithFields(c).Error(msg)
@@ -34,15 +46,21 @@ func (s *shim) Debug(msg string, c map[string]interface{}) {
 	s.e.WithFields(c).Debug(msg)
 }
 func (s *shim) DebugEnabled() bool {
-	lvl := logrus.GetLevel()
+	lvl := s.l.GetLevel()
 	return lvl >= logrus.DebugLevel
 }
 
 // StandardLogger returns a newrelic.Logger which forwards agent log messages to
 // the logrus package-level exported logger.
 func StandardLogger() newrelic.Logger {
+	return Transform(logrus.StandardLogger())
+}
+
+// Transform turns a *logrus.Logger into a newrelic.Logger.
+func Transform(l *logrus.Logger) newrelic.Logger {
 	return &shim{
-		e: logrus.WithFields(logrus.Fields{
+		l: l,
+		e: l.WithFields(logrus.Fields{
 			"component": "newrelic",
 		}),
 	}